Battery decode arithmetic: five fixes with named evidence, three pins
Evidence first, per the row - every change names what settles it, and
what nothing settles is pinned as current behaviour and left alone
(a review's withdrawn BMW-PHEV claim is the cautionary case).

FIXED, each with a fails-before test:
- RANGE-ROVER-PHEV: the 24-bit current assembled both high bytes with
  << 8, colliding them in one lane. Evidence: the driver's own
  declaration (0 - 16777215, offset -209715.175) - a 24-bit range whose
  midpoint only a 24-bit assembly reaches.
- TESLA-LEGACY: brick temperatures pass through uint8_t intermediates
  while their own decode is (raw * 0.5) - 40, range -40..+87.5 C - a
  subzero pack wrapped to ~+215 C. Evidence: C++ semantics.
- IMIEV: temperature channels 2 and 3 read u8[1] while guarding on
  u8[2]/u8[3] - the guards name the intended bytes. Evidence: the
  driver's own guards. Cell voltages truncated (3.7f * 1000 = 3699);
  now lroundf. Evidence: float semantics. The two 88-entry instance
  arrays were also uninitialized heap reads until every sensor reported;
  zero-initialized.
- RELION-LV: min_cell_temperature was decoded and never used (both
  datalayer fields carried the max), and its decode read u8[2] - an id
  byte whose value would be a constant -49 C per the case's own captured
  frame 47 01 01 47 01 01, two (value, id, id) triplets in the same
  shape the driver's ID3 uses for voltages. Min now wired, from u8[3].
- ENNOID: the only reported temperature landed in the MIN field with
  the max one degree COLDER, both in raw C against dC fields. Evidence:
  the datalayer's documented unit and the max >= min invariant. The
  author's synthetic 1-degree spread is kept, ordered and scaled.

PINNED as current behaviour, not changed:
- TESLA-LEGACY hwID 79/89: labelled 100kWh, sets 70000 Wh. One of the
  two is wrong (5 of 6 groups have comment == value) but nothing in the
  tree, the originating commit, or public pack tables says which.
- VOLVO-SPA-HYBRID 0x369: low byte from u8[6] (self-referential) - and
  the SIBLING driver has the identical expression, so no sibling settles
  the layout; the field is display-only.
- RELION-LV 0x264: discharge current reads the regen bytes; its only
  consumer is commented out, and no trace names the discharge bytes.

Seven revert-mutations each fail exactly their named test. Host suite
215/215 + shuffle seeds 1/7/42/1234/99999; lilygo_330 firmware builds;
clang-format clean.

+++ Software/src/battery/TESLA-LEGACY-BATTERY.h
@@ -83,8 +83,11 @@ class TeslaLegacyBattery : public CanBattery {
   uint16_t battery_BrickVoltageMin = 0;
   uint8_t battery_BrickTempMaxNum = 0;
   uint8_t battery_BrickTempMinNum = 0;
-  uint8_t battery_BrickModelTMax = 0;
-  uint8_t battery_BrickModelTMin = 0;
+  // int16_t, not uint8_t (wq311): the decode is (raw * 0.5) - 40, whose own
+  // range is -40..+87.5 C - a subzero pack wrapped to ~+215 C and the datalayer
+  // then showed ~2150 dC.
+  int16_t battery_BrickModelTMax = 0;
+  int16_t battery_BrickModelTMin = 0;
   uint8_t battery_BrickVoltageMaxNum = 0;
   uint8_t battery_BrickVoltageMinNum = 0;
   //0x5D2
